Move hybrid decode surface bindings into ANERuntime
The split-hybrid and fused-hybrid trunk kernels defined their surface
layouts, but the binding knowledge lived one module over in Espresso:
HybridDecodeSurfaceHandles and FusedHybridDecodeSurfaceHandles encoded
positional surface indices, rebind choreography, fp16 cache byte math,
and the fused/split ffnOut selection inside a consumer module.

Both handle types move to ANERuntime beside the kernel sets that own
the layouts. Consumers keep compiling unchanged; the seam now reads
the right way around: kernel sets expose bound stages, not raw parts.

Also collapses the post-attention eval branch: HybridDecodeKernelSet
gains evalPostAttention() and postAttentionStageName so the four call
sites in DecodeForwardPass stop branching on usesFusedPostAttention
and hand-rolling the stage label for error messages.

mapSurfaceIOToANEError moves to ANERuntime (public) next to ANEError;
Espresso's copy is deleted rather than duplicated per module.
